Fix flag selection logic, resolve #261, #258, #257.

+++ ccs_workspace/qc15_common/qc15.h
@@ -116,7 +116,7 @@ typedef struct {
     uint8_t badges_seen[57];
     uint8_t badges_downloaded[57];
     uint8_t badges_uploaded[57];
-    uint16_t flag_unlocks;
+    uint32_t flag_unlocks;
     uint32_t ubers_seen, ubers_downloaded, ubers_uploaded;
     uint32_t handlers_seen, handlers_downloaded, handlers_uploaded;
     uint8_t file_lights_on : 1;
